requirements.txt
facenet-pytorch
scikit-learn
numpy
pillow
opencv-python
tqdm
joblib
streamlit